requirements.txt
# pip instal -r requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-core==0.3.12
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.137
Levenshtein==0.26.0
mando==0.7.1
multidict==6.1.0
numpy==1.26.4
ollama==0.3.3
orjson==3.10.10
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-Levenshtein==0.26.0
PyYAML==6.0.2
radon==6.0.1
RapidFuzz==3.10.0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
